feat: finish build in playground
