+++ package.json
@@ -7,7 +7,8 @@
     "dev": "npm run start:client & npm run start:node",
     "start:client": "cd src/client && tsc -w",
     "start:node": "cd src/node && tsc -w",
-    "dev:island": "node ./bin/island"
+    "dev:playground": "node ./bin/island playground",
+    "build:playground": "node ./bin/island build playground"
   },
   "keywords": [],
   "author": "",
@@ -25,10 +26,13 @@
     "vite": "^3.0.9"
   },
   "devDependencies": {
+    "@types/fs-extra": "^9.0.13",
     "@types/koa": "^2.13.5",
     "@types/koa-router": "^7.4.4",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
+    "ora": "^6.1.2",
+    "rollup": "^2.78.1",
     "tsx": "^3.8.2",
     "typescript": "^4.7.4"
   }
