+++ src/client/app/ssr-entry.tsx
@@ -2,6 +2,6 @@ import { renderToString } from 'react-dom/server';
 import { App } from './app';
 
 // For ssr component render
-export async function render() {
+export function render() {
   return renderToString(<App />);
 }
