+++ src/node/constants/index.ts
@@ -1,12 +1,21 @@
 import { join } from 'path';
 
 export const isProduction = () => process.env.NODE_ENV === 'production';
-export const CLIENT_APP_PATH = join(
+export const CLIENT_ENTRY_PATH = join(
   __dirname,
   '../../../dist/client/app/client-entry.js'
+);
+
+export const SERVER_ENTRY_PATH = join(
+  __dirname,
+  '../../../dist/client/app/ssr-entry.js'
 );
 
 export const THEME_PATH = join(
   __dirname,
   '../../../dist/client/theme/index.js'
 );
+
+export const TEMP_PATH = 'node_modules/.island';
+
+export const SERVER_OUTPUT_PATH = join(TEMP_PATH, 'ssr-entry.mjs');
