+++ src/node/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "target": "ES2022",
     "module": "commonjs",
     "outDir": "../../dist/node"
   }
