+++ src/node/utils.ts
@@ -1,0 +1,4 @@
+export const dynamicImport = new Function(
+  'modulePath',
+  'return import(modulePath)'
+);
